Add makefile rule for generating gcode.
This is a minimal config for our Tinkrine Ditto+ printer,
just the values that differ from the defaults.

Note that 220C is the recommended print temperature for
the recent tinkerine brand filament we're using, but that's
considerably hotter than the 200C default I've been using
so some verification may be needed. Could definitely use
more viscosity than we've been getting with the first layers
at 205C though.