Move mounting stud.
Generalize the offset to a variable so it can be shared between
the surround and the through-hole.

I remembered the dimension correctly, but forgot to account
for the extra sqrt(2) for it being at 45 degrees relative
to the other mounting holes.

NB this breaks printability because it's outside the defined
size envelope so the mounting holes aren't connected to the
main ring. I need to update the connecting shapes.

Also fix a dimension issue with the updated panel thickness.